refactor: Change directory structure (#4)
