+++ Server/src/main/java/com/gdscswu_server/server/global/config/JpaConfig.java
@@ -1,8 +1,9 @@
-package com.gdscswu_server.server.config;
+package com.gdscswu_server.server.global.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @Configuration
 @EnableJpaAuditing // JPA Auditing 활성화
-public class JpaConfig {}
+public class JpaConfig {
+}
