+++ Server/src/main/java/com/gdscswu_server/server/global/util/auth/dto/OAuthAttributes.java
@@ -1,4 +1,4 @@
-package com.gdscswu_server.server.config.auth.dto;
+package com.gdscswu_server.server.global.util.auth.dto;
 
 import com.gdscswu_server.server.domain.member.domain.Member;
 import lombok.Builder;
@@ -12,19 +12,19 @@ public class OAuthAttributes {
     private String nameAttributeKey;
     private String name;
     private String email;
-    private String profile_image;
+    private String profileImagePath;
 
     @Builder
-    public OAuthAttributes(Map<String, Object> attributes, String nameAttributeKey, String name, String email, String profile_image) {
+    public OAuthAttributes(Map<String, Object> attributes, String nameAttributeKey, String name, String email, String profileImagePath) {
         this.attributes = attributes;
         this.nameAttributeKey = nameAttributeKey;
         this.name = name;
         this.email = email;
-        this.profile_image = profile_image;
+        this.profileImagePath = profileImagePath;
     }
 
     public static OAuthAttributes of(String registrationId, String userNameAttributeName, Map<String, Object> attributes) {
-        if("google".equals(registrationId)) {
+        if ("google".equals(registrationId)) {
             return ofGoogle("id", attributes);
         }
 
@@ -35,7 +35,7 @@ public class OAuthAttributes {
         return OAuthAttributes.builder()
                 .name((String) attributes.get("name"))
                 .email((String) attributes.get("email"))
-                .profile_image((String) attributes.get("profile_image"))
+                .profileImagePath((String) attributes.get("profile_image"))
                 .attributes(attributes)
                 .nameAttributeKey(userNameAttributeName)
                 .build();
@@ -47,7 +47,7 @@ public class OAuthAttributes {
         return OAuthAttributes.builder()
                 .name((String) response.get("name"))
                 .email((String) response.get("email"))
-                .profile_image((String) response.get("profile_image"))
+                .profileImagePath((String) response.get("profile_image"))
                 .attributes(response)
                 .nameAttributeKey(userNameAttributeName)
                 .build();
@@ -56,8 +56,8 @@ public class OAuthAttributes {
     public Member toEntity() {
         return Member.builder()
                 .name(name)
-                .email(email)
-                .profile_image(profile_image.getBytes())
+                .googleEmail(email)
+                .profileImagePath(profileImagePath)
                 .build();
     }
 }
